I remake maven project, starting from zero, now its work corectly, build jar with dependency and run it corectly I think
