+++ myproject-bl/pom.xml
@@ -5,9 +5,10 @@
     <modelVersion>4.0.0</modelVersion>
     <packaging>jar</packaging>
 
-    <!--<groupId>clouway</groupId>-->
+    <groupId>clouway</groupId>
     <artifactId>myproject-bl</artifactId>
     <version>1.0-SNAPSHOT</version>
+
     <parent>
         <groupId>clouway</groupId>
         <artifactId>buildtools</artifactId>
