+++ myproject-bl/src/main/java/package2/ProjectBL.java
@@ -1,0 +1,10 @@
+package package2;
+
+import org.apache.log4j.Logger;
+
+public class ProjectBL {
+    public void print() {
+        Logger log = Logger.getLogger(ProjectBL.class);
+        log.info("1 more user in the system");
+    }
+}
