+++ myproject-bl/src/main/resources/log4j.properties
