--- myproject-bl/src/myprojectbl/BLClass.java
@@ -1,9 +1,0 @@
-package myprojectbl;
-
-import org.apache.log4j.Logger;
-
-public class BLClass {
-    public void print() {
-        System.out.println("you are in BLClass");
-    }
-}
