--- myproject-bo/src/projectBO/ProjectBO.java
@@ -1,9 +1,0 @@
-package projectBO;
-
-import org.apache.log4j.Logger;
-
-public class ProjectBO {
-    public void print() {
-        System.out.println("you are in ProjectBO");
-    }
-}
