+++ myproject-pl/pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
     <packaging>jar</packaging>
 
-    <!--<groupId>clouway</groupId>-->
+    <groupId>clouway</groupId>
     <artifactId>myproject-pl</artifactId>
     <version>1.0-SNAPSHOT</version>
