+++ myproject-pl/src/main/java/package4/ProjectPL.java
@@ -1,0 +1,10 @@
+package package4;
+
+import org.apache.log4j.Logger;
+
+public class ProjectPL {
+    public void print() {
+        Logger log = Logger.getLogger(ProjectPL.class);
+        log.info("1 more user in the system");
+    }
+}
