--- myproject-pl/src/projectPL/ProjectPL.java
@@ -1,10 +1,0 @@
-package projectPL;
-
-import org.apache.log4j.Logger;
-import com.mysql.jdbc.*;
-
-public class ProjectPL {
-    public void print() {
-        System.out.println("you are in ProjectPL");
-    }
-}
