+++ myproject/pom.xml
@@ -9,12 +9,6 @@
     <artifactId>myproject</artifactId>
     <version>1.0-SNAPSHOT</version>
 
-    <parent>
-        <groupId>clouway</groupId>
-        <artifactId>buildtools</artifactId>
-        <version>1.0-SNAPSHOT</version>
-    </parent>
-
     <dependencies>
         <dependency>
             <groupId>clouway</groupId>
@@ -32,5 +26,11 @@
             <version>1.0-SNAPSHOT</version>
         </dependency>
     </dependencies>
-    
+
+    <parent>
+        <groupId>clouway</groupId>
+        <artifactId>buildtools</artifactId>
+        <version>1.0-SNAPSHOT</version>
+    </parent>
+
 </project>
